explorer - consolidate upload/download code and provide a "Upload..." action (web)
cc @isidorn
