+++ src/vs/workbench/contrib/files/browser/views/explorerViewer.ts
@@ -7,9 +7,9 @@ import { IListAccessibilityProvider } from 'vs/base/browser/ui/list/listWidget';
 import * as DOM from 'vs/base/browser/dom';
 import * as glob from 'vs/base/common/glob';
 import { IListVirtualDelegate, ListDragOverEffect } from 'vs/base/browser/ui/list/list';
-import { IProgressService, ProgressLocation, IProgressStep, IProgress } from 'vs/platform/progress/common/progress';
+import { IProgressService, ProgressLocation, } from 'vs/platform/progress/common/progress';
 import { INotificationService, Severity } from 'vs/platform/notification/common/notification';
-import { IFileService, FileKind, FileOperationError, FileOperationResult, FileSystemProviderCapabilities, ByteSize } from 'vs/platform/files/common/files';
+import { IFileService, FileKind, FileOperationError, FileOperationResult } from 'vs/platform/files/common/files';
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
 import { IWorkspaceContextService, WorkbenchState } from 'vs/platform/workspace/common/workspace';
 import { IDisposable, Disposable, dispose, toDisposable, DisposableStore } from 'vs/base/common/lifecycle';
@@ -19,8 +19,8 @@ import { ITreeNode, ITreeFilter, TreeVisibility, IAsyncDataSource, ITreeSorter, 
 import { IContextViewService } from 'vs/platform/contextview/browser/contextView';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IFilesConfiguration, VIEW_ID } from 'vs/workbench/contrib/files/common/files';
-import { dirname, joinPath, basename, distinctParents } from 'vs/base/common/resources';
+import { IFilesConfiguration } from 'vs/workbench/contrib/files/common/files';
+import { dirname, joinPath, distinctParents } from 'vs/base/common/resources';
 import { InputBox, MessageType } from 'vs/base/browser/ui/inputbox/inputBox';
 import { localize } from 'vs/nls';
 import { attachInputBoxStyler } from 'vs/platform/theme/common/styler';
@@ -30,17 +30,15 @@ import { equals, deepClone } from 'vs/base/common/objects';
 import * as path from 'vs/base/common/path';
 import { ExplorerItem, NewExplorerItem } from 'vs/workbench/contrib/files/common/explorerModel';
 import { compareFileExtensionsDefault, compareFileNamesDefault, compareFileNamesUpper, compareFileExtensionsUpper, compareFileNamesLower, compareFileExtensionsLower, compareFileNamesUnicode, compareFileExtensionsUnicode } from 'vs/base/common/comparers';
-import { fillResourceDataTransfers, CodeDataTransfers, extractResources, containsDragType } from 'vs/workbench/browser/dnd';
+import { fillResourceDataTransfers, CodeDataTransfers, containsDragType } from 'vs/workbench/browser/dnd';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IDragAndDropData, DataTransfers } from 'vs/base/browser/dnd';
 import { Schemas } from 'vs/base/common/network';
 import { NativeDragAndDropData, ExternalElementsDragAndDropData, ElementsDragAndDropData } from 'vs/base/browser/ui/list/listView';
 import { isMacintosh, isWeb } from 'vs/base/common/platform';
-import { IDialogService, IConfirmation, getFileNamesMessage } from 'vs/platform/dialogs/common/dialogs';
-import { IHostService } from 'vs/workbench/services/host/browser/host';
+import { IDialogService, getFileNamesMessage } from 'vs/platform/dialogs/common/dialogs';
 import { IWorkspaceEditingService } from 'vs/workbench/services/workspaces/common/workspaceEditing';
 import { URI } from 'vs/base/common/uri';
-import { RunOnceWorker } from 'vs/base/common/async';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IWorkspaceFolderCreationData } from 'vs/platform/workspaces/common/workspaces';
 import { findValidPasteFileTarget } from 'vs/workbench/contrib/files/browser/fileActions';
@@ -49,16 +47,16 @@ import { Emitter, Event, EventMultiplexer } from 'vs/base/common/event';
 import { ITreeCompressionDelegate } from 'vs/base/browser/ui/tree/asyncDataTree';
 import { ICompressibleTreeRenderer } from 'vs/base/browser/ui/tree/objectTree';
 import { ICompressedTreeNode } from 'vs/base/browser/ui/tree/compressedObjectTreeModel';
-import { VSBuffer, newWriteableBufferStream } from 'vs/base/common/buffer';
 import { ILabelService } from 'vs/platform/label/common/label';
 import { isNumber } from 'vs/base/common/types';
 import { domEvent } from 'vs/base/browser/event';
 import { IEditableData } from 'vs/workbench/common/views';
 import { IEditorInput } from 'vs/workbench/common/editor';
-import { CancellationTokenSource, CancellationToken } from 'vs/base/common/cancellation';
 import { IUriIdentityService } from 'vs/workbench/services/uriIdentity/common/uriIdentity';
 import { ResourceFileEdit } from 'vs/editor/browser/services/bulkEditService';
 import { IExplorerService } from 'vs/workbench/contrib/files/browser/files';
+import { BrowserFileUpload, NativeFileImport, getMultipleFilesOverwriteConfirm } from 'vs/workbench/contrib/files/browser/fileImportExport';
+import { toErrorMessage } from 'vs/base/common/errorMessage';
 
 export class ExplorerDelegate implements IListVirtualDelegate<ExplorerItem> {
 
@@ -754,59 +752,6 @@ export class FileSorter implements ITreeSorter<ExplorerItem> {
 	}
 }
 
-function getFileOverwriteConfirm(name: string): IConfirmation {
-	return {
-		message: localize('confirmOverwrite', "A file or folder with the name '{0}' already exists in the destination folder. Do you want to replace it?", name),
-		detail: localize('irreversible', "This action is irreversible!"),
-		primaryButton: localize({ key: 'replaceButtonLabel', comment: ['&& denotes a mnemonic'] }, "&&Replace"),
-		type: 'warning'
-	};
-}
-
-function getMultipleFilesOverwriteConfirm(files: URI[]): IConfirmation {
-	if (files.length > 1) {
-		return {
-			message: localize('confirmManyOverwrites', "The following {0} files and/or folders already exist in the destination folder. Do you want to replace them?", files.length),
-			detail: getFileNamesMessage(files) + '\n' + localize('irreversible', "This action is irreversible!"),
-			primaryButton: localize({ key: 'replaceButtonLabel', comment: ['&& denotes a mnemonic'] }, "&&Replace"),
-			type: 'warning'
-		};
-	}
-
-	return getFileOverwriteConfirm(basename(files[0]));
-}
-
-interface IWebkitDataTransfer {
-	items: IWebkitDataTransferItem[];
-}
-
-interface IWebkitDataTransferItem {
-	webkitGetAsEntry(): IWebkitDataTransferItemEntry;
-}
-
-interface IWebkitDataTransferItemEntry {
-	name: string | undefined;
-	isFile: boolean;
-	isDirectory: boolean;
-
-	file(resolve: (file: File) => void, reject: () => void): void;
-	createReader(): IWebkitDataTransferItemEntryReader;
-}
-
-interface IWebkitDataTransferItemEntryReader {
-	readEntries(resolve: (file: IWebkitDataTransferItemEntry[]) => void, reject: () => void): void
-}
-
-interface IUploadOperation {
-	startTime: number;
-	progressScheduler: RunOnceWorker<IProgressStep>;
-
-	filesTotal: number;
-	filesUploaded: number;
-
-	totalBytesUploaded: number;
-}
-
 export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 	private static readonly CONFIRM_DND_SETTING_KEY = 'explorer.confirmDragAndDrop';
 
@@ -817,7 +762,6 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 	private dropEnabled = false;
 
 	constructor(
-		@INotificationService private notificationService: INotificationService,
 		@IExplorerService private explorerService: IExplorerService,
 		@IEditorService private editorService: IEditorService,
 		@IDialogService private dialogService: IDialogService,
@@ -825,9 +769,7 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 		@IFileService private fileService: IFileService,
 		@IConfigurationService private configurationService: IConfigurationService,
 		@IInstantiationService private instantiationService: IInstantiationService,
-		@IHostService private hostService: IHostService,
 		@IWorkspaceEditingService private workspaceEditingService: IWorkspaceEditingService,
-		@IProgressService private readonly progressService: IProgressService,
 		@IUriIdentityService private readonly uriIdentityService: IUriIdentityService
 	) {
 		this.toDispose = [];
@@ -1024,358 +966,25 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 			return;
 		}
 
-		// Desktop DND (Import file)
-		if (data instanceof NativeDragAndDropData) {
-			const cts = new CancellationTokenSource();
-
-			if (isWeb) {
-				// Indicate progress globally
-				const dropPromise = this.progressService.withProgress({
-					location: ProgressLocation.Window,
-					delay: 800,
-					cancellable: true,
-					title: localize('uploadingFiles', "Uploading")
-				}, async progress => {
-					try {
-						await this.handleWebExternalDrop(resolvedTarget, originalEvent, progress, cts.token);
-					} catch (error) {
-						this.notificationService.warn(error);
-					}
-				}, () => cts.dispose(true));
-				// Also indicate progress in the files view
-				this.progressService.withProgress({ location: VIEW_ID, delay: 500 }, () => dropPromise);
-			} else {
-				try {
-					await this.handleExternalDrop(resolvedTarget, originalEvent, cts.token);
-				} catch (error) {
-					this.notificationService.warn(error);
-				}
-			}
-		}
-		// In-Explorer DND (Move/Copy file)
-		else {
-			this.handleExplorerDrop(data as ElementsDragAndDropData<ExplorerItem, ExplorerItem[]>, resolvedTarget, originalEvent).then(undefined, e => this.notificationService.warn(e));
-		}
-	}
-
-	private async handleWebExternalDrop(target: ExplorerItem, originalEvent: DragEvent, progress: IProgress<IProgressStep>, token: CancellationToken): Promise<void> {
-		const items = (originalEvent.dataTransfer as unknown as IWebkitDataTransfer).items;
-
-		// Somehow the items thing is being modified at random, maybe as a security
-		// measure since this is a DND operation. As such, we copy the items into
-		// an array we own as early as possible before using it.
-		const entries: IWebkitDataTransferItemEntry[] = [];
-		for (const item of items) {
-			entries.push(item.webkitGetAsEntry());
-		}
-
-		const results: { isFile: boolean, resource: URI }[] = [];
-		const operation: IUploadOperation = {
-			startTime: Date.now(),
-			progressScheduler: new RunOnceWorker<IProgressStep>(steps => { progress.report(steps[steps.length - 1]); }, 1000),
-
-			filesTotal: entries.length,
-			filesUploaded: 0,
-
-			totalBytesUploaded: 0
-		};
-
-		for (let entry of entries) {
-			if (token.isCancellationRequested) {
-				break;
-			}
-
-			// Confirm overwrite as needed
-			if (target && entry.name && target.getChild(entry.name)) {
-				const { confirmed } = await this.dialogService.confirm(getFileOverwriteConfirm(entry.name));
-				if (!confirmed) {
-					continue;
-				}
-
-				await this.explorerService.applyBulkEdit([new ResourceFileEdit(joinPath(target.resource, entry.name), undefined, { recursive: true })], {
-					undoLabel: localize('overwrite', "Overwrite {0}", entry.name),
-					progressLabel: localize('overwriting', "Overwriting {0}", entry.name),
-				});
-
-				if (token.isCancellationRequested) {
-					break;
-				}
-			}
-
-			// Upload entry
-			const result = await this.doUploadWebFileEntry(entry, target.resource, target, progress, operation, token);
-			if (result) {
-				results.push(result);
-			}
-		}
-
-		operation.progressScheduler.dispose();
-
-		// Open uploaded file in editor only if we upload just one
-		const firstUploadedFile = results[0];
-		if (!token.isCancellationRequested && firstUploadedFile?.isFile) {
-			await this.editorService.openEditor({ resource: firstUploadedFile.resource, options: { pinned: true } });
-		}
-	}
-
-	private async doUploadWebFileEntry(entry: IWebkitDataTransferItemEntry, parentResource: URI, target: ExplorerItem | undefined, progress: IProgress<IProgressStep>, operation: IUploadOperation, token: CancellationToken): Promise<{ isFile: boolean, resource: URI } | undefined> {
-		if (token.isCancellationRequested || !entry.name || (!entry.isFile && !entry.isDirectory)) {
-			return undefined;
-		}
-
-		// Report progress
-		let fileBytesUploaded = 0;
-		const reportProgress = (fileSize: number, bytesUploaded: number): void => {
-			fileBytesUploaded += bytesUploaded;
-			operation.totalBytesUploaded += bytesUploaded;
-
-			const bytesUploadedPerSecond = operation.totalBytesUploaded / ((Date.now() - operation.startTime) / 1000);
-
-			// Small file
-			let message: string;
-			if (fileSize < ByteSize.MB) {
-				if (operation.filesTotal === 1) {
-					message = `${entry.name}`;
-				} else {
-					message = localize('uploadProgressSmallMany', "{0} of {1} files ({2}/s)", operation.filesUploaded, operation.filesTotal, ByteSize.formatSize(bytesUploadedPerSecond));
-				}
-			}
-
-			// Large file
-			else {
-				message = localize('uploadProgressLarge', "{0} ({1} of {2}, {3}/s)", entry.name, ByteSize.formatSize(fileBytesUploaded), ByteSize.formatSize(fileSize), ByteSize.formatSize(bytesUploadedPerSecond));
-			}
-
-			// Report progress but limit to update only once per second
-			operation.progressScheduler.work({ message });
-		};
-		operation.filesUploaded++;
-		reportProgress(0, 0);
-
-		// Handle file upload
-		const resource = joinPath(parentResource, entry.name);
-		if (entry.isFile) {
-			const file = await new Promise<File>((resolve, reject) => entry.file(resolve, reject));
-
-			if (token.isCancellationRequested) {
-				return undefined;
-			}
-
-			// Chrome/Edge/Firefox support stream method, but only use it for
-			// larger files to reduce the overhead of the streaming approach
-			if (typeof file.stream === 'function' && file.size > ByteSize.MB) {
-				await this.doUploadWebFileEntryBuffered(resource, file, reportProgress, token);
-			}
-
-			// Fallback to unbuffered upload for other browsers or small files
-			else {
-				await this.doUploadWebFileEntryUnbuffered(resource, file, reportProgress);
-			}
-
-			return { isFile: true, resource };
-		}
-
-		// Handle folder upload
-		else {
-
-			// Create target folder
-			await this.fileService.createFolder(resource);
-
-			if (token.isCancellationRequested) {
-				return undefined;
-			}
-
-			// Recursive upload files in this directory
-			const dirReader = entry.createReader();
-			const childEntries: IWebkitDataTransferItemEntry[] = [];
-			let done = false;
-			do {
-				const childEntriesChunk = await new Promise<IWebkitDataTransferItemEntry[]>((resolve, reject) => dirReader.readEntries(resolve, reject));
-				if (childEntriesChunk.length > 0) {
-					childEntries.push(...childEntriesChunk);
-				} else {
-					done = true; // an empty array is a signal that all entries have been read
-				}
-			} while (!done && !token.isCancellationRequested);
-
-			// Update operation total based on new counts
-			operation.filesTotal += childEntries.length;
-
-			// Upload all entries as files to target
-			const folderTarget = target && target.getChild(entry.name) || undefined;
-			for (let childEntry of childEntries) {
-				await this.doUploadWebFileEntry(childEntry, resource, folderTarget, progress, operation, token);
-			}
-
-			return { isFile: false, resource };
-		}
-	}
-
-	private async doUploadWebFileEntryBuffered(resource: URI, file: File, progressReporter: (fileSize: number, bytesUploaded: number) => void, token: CancellationToken): Promise<void> {
-		const writeableStream = newWriteableBufferStream({
-			// Set a highWaterMark to prevent the stream
-			// for file upload to produce large buffers
-			// in-memory
-			highWaterMark: 10
-		});
-		const writeFilePromise = this.fileService.writeFile(resource, writeableStream);
-
-		// Read the file in chunks using File.stream() web APIs
 		try {
-			const reader: ReadableStreamDefaultReader<Uint8Array> = file.stream().getReader();
 
-			let res = await reader.read();
-			while (!res.done) {
-				if (token.isCancellationRequested) {
-					return undefined;
+			// Desktop DND (Import file)
+			if (data instanceof NativeDragAndDropData) {
+				if (isWeb) {
+					const browserUpload = this.instantiationService.createInstance(BrowserFileUpload);
+					await browserUpload.upload(target, originalEvent);
+				} else {
+					const nativeImport = this.instantiationService.createInstance(NativeFileImport);
+					await nativeImport.import(resolvedTarget, originalEvent);
 				}
-
-				// Write buffer into stream but make sure to wait
-				// in case the highWaterMark is reached
-				const buffer = VSBuffer.wrap(res.value);
-				await writeableStream.write(buffer);
-
-				if (token.isCancellationRequested) {
-					return undefined;
-				}
-
-				// Report progress
-				progressReporter(file.size, buffer.byteLength);
-
-				res = await reader.read();
 			}
-			writeableStream.end(undefined);
+
+			// In-Explorer DND (Move/Copy file)
+			else {
+				await this.handleExplorerDrop(data as ElementsDragAndDropData<ExplorerItem, ExplorerItem[]>, resolvedTarget, originalEvent);
+			}
 		} catch (error) {
-			writeableStream.error(error);
-			writeableStream.end();
-		}
-
-		if (token.isCancellationRequested) {
-			return undefined;
-		}
-
-		// Wait for file being written to target
-		await writeFilePromise;
-	}
-
-	private doUploadWebFileEntryUnbuffered(resource: URI, file: File, progressReporter: (fileSize: number, bytesUploaded: number) => void): Promise<void> {
-		return new Promise<void>((resolve, reject) => {
-			const reader = new FileReader();
-			reader.onload = async event => {
-				try {
-					if (event.target?.result instanceof ArrayBuffer) {
-						const buffer = VSBuffer.wrap(new Uint8Array(event.target.result));
-						await this.fileService.writeFile(resource, buffer);
-
-						// Report progress
-						progressReporter(file.size, buffer.byteLength);
-					} else {
-						throw new Error('Could not read from dropped file.');
-					}
-
-					resolve();
-				} catch (error) {
-					reject(error);
-				}
-			};
-
-			// Start reading the file to trigger `onload`
-			reader.readAsArrayBuffer(file);
-		});
-	}
-
-	private async handleExternalDrop(target: ExplorerItem, originalEvent: DragEvent, token: CancellationToken): Promise<void> {
-
-		// Check for dropped external files to be folders
-		const droppedResources = extractResources(originalEvent, true);
-		const result = await this.fileService.resolveAll(droppedResources.map(droppedResource => ({ resource: droppedResource.resource })));
-
-		if (token.isCancellationRequested) {
-			return;
-		}
-
-		// Pass focus to window
-		this.hostService.focus();
-
-		// Handle folders by adding to workspace if we are in workspace context and if dropped on top
-		const folders = result.filter(r => r.success && r.stat && r.stat.isDirectory).map(result => ({ uri: result.stat!.resource }));
-		if (folders.length > 0 && target.isRoot) {
-			const buttons = [
-				folders.length > 1 ? localize('copyFolders', "&&Copy Folders") : localize('copyFolder', "&&Copy Folder"),
-				localize('cancel', "Cancel")
-			];
-			const workspaceFolderSchemas = this.contextService.getWorkspace().folders.map(f => f.uri.scheme);
-			let message = folders.length > 1 ? localize('copyfolders', "Are you sure to want to copy folders?") : localize('copyfolder', "Are you sure to want to copy '{0}'?", basename(folders[0].uri));
-			if (folders.some(f => workspaceFolderSchemas.indexOf(f.uri.scheme) >= 0)) {
-				// We only allow to add a folder to the workspace if there is already a workspace folder with that scheme
-				buttons.unshift(folders.length > 1 ? localize('addFolders', "&&Add Folders to Workspace") : localize('addFolder', "&&Add Folder to Workspace"));
-				message = folders.length > 1 ? localize('dropFolders', "Do you want to copy the folders or add the folders to the workspace?")
-					: localize('dropFolder', "Do you want to copy '{0}' or add '{0}' as a folder to the workspace?", basename(folders[0].uri));
-			}
-
-			const { choice } = await this.dialogService.show(Severity.Info, message, buttons);
-			if (choice === buttons.length - 3) {
-				return this.workspaceEditingService.addFolders(folders);
-			}
-			if (choice === buttons.length - 2) {
-				return this.addResources(target, droppedResources.map(res => res.resource), token);
-			}
-
-			return undefined;
-		}
-
-		// Handle dropped files (only support FileStat as target)
-		else if (target instanceof ExplorerItem) {
-			return this.addResources(target, droppedResources.map(res => res.resource), token);
-		}
-	}
-
-	private async addResources(target: ExplorerItem, resources: URI[], token: CancellationToken): Promise<void> {
-		if (resources && resources.length > 0) {
-
-			// Resolve target to check for name collisions and ask user
-			const targetStat = await this.fileService.resolve(target.resource);
-
-			if (token.isCancellationRequested) {
-				return;
-			}
-
-			// Check for name collisions
-			const targetNames = new Set<string>();
-			const caseSensitive = this.fileService.hasCapability(target.resource, FileSystemProviderCapabilities.PathCaseSensitive);
-			if (targetStat.children) {
-				targetStat.children.forEach(child => {
-					targetNames.add(caseSensitive ? child.name : child.name.toLowerCase());
-				});
-			}
-
-			const resourcesFiltered = (await Promise.all(resources.map(async resource => {
-				if (targetNames.has(caseSensitive ? basename(resource) : basename(resource).toLowerCase())) {
-					const confirmationResult = await this.dialogService.confirm(getFileOverwriteConfirm(basename(resource)));
-					if (!confirmationResult.confirmed) {
-						return undefined;
-					}
-				}
-				return resource;
-			}))).filter(r => r instanceof URI) as URI[];
-			const resourceFileEdits = resourcesFiltered.map(resource => {
-				const sourceFileName = basename(resource);
-				const targetFile = joinPath(target.resource, sourceFileName);
-				return new ResourceFileEdit(resource, targetFile, { overwrite: true, copy: true });
-			});
-
-			await this.explorerService.applyBulkEdit(resourceFileEdits, {
-				undoLabel: resourcesFiltered.length === 1 ? localize('copyFile', "Copy {0}", basename(resourcesFiltered[0])) : localize('copynFile', "Copy {0} resources", resourcesFiltered.length),
-				progressLabel: resourcesFiltered.length === 1 ? localize('copyingFile', "Copying {0}", basename(resourcesFiltered[0])) : localize('copyingnFile', "Copying {0} resources", resourcesFiltered.length)
-			});
-
-			// if we only add one file, just open it directly
-			if (resourceFileEdits.length === 1) {
-				const item = this.explorerService.findClosest(resourceFileEdits[0].newResource!);
-				if (item && !item.isDirectory) {
-					this.editorService.openEditor({ resource: item.resource, options: { pinned: true } });
-				}
-			}
+			this.dialogService.show(Severity.Error, toErrorMessage(error), [localize('ok', 'OK')]);
 		}
 	}
 
@@ -1417,15 +1026,15 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 
 		const sources = items.filter(s => !s.isRoot);
 		if (isCopy) {
-			await this.doHandleExplorerDropOnCopy(sources, target);
-		} else {
-			return this.doHandleExplorerDropOnMove(sources, target);
+			return this.doHandleExplorerDropOnCopy(sources, target);
 		}
+
+		return this.doHandleExplorerDropOnMove(sources, target);
 	}
 
-	private doHandleRootDrop(roots: ExplorerItem[], target: ExplorerItem): Promise<void> {
+	private async doHandleRootDrop(roots: ExplorerItem[], target: ExplorerItem): Promise<void> {
 		if (roots.length === 0) {
-			return Promise.resolve(undefined);
+			return;
 		}
 
 		const folders = this.contextService.getWorkspace().folders;
@@ -1453,10 +1062,12 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 		}
 
 		workspaceCreationData.splice(targetIndex, 0, ...rootsToMove);
+
 		return this.workspaceEditingService.updateFolders(0, workspaceCreationData.length, workspaceCreationData);
 	}
 
 	private async doHandleExplorerDropOnCopy(sources: ExplorerItem[], target: ExplorerItem): Promise<void> {
+
 		// Reuse duplicate action when user copies
 		const incrementalNaming = this.configurationService.getValue<IFilesConfiguration>().explorer.incrementalNaming;
 		const resourceFileEdits = sources.map(({ resource, isDirectory }) => (new ResourceFileEdit(resource, findValidPasteFileTarget(this.explorerService, target, { resource, isDirectory, allowOverwrite: false }, incrementalNaming), { copy: true })));
@@ -1487,6 +1098,7 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 		try {
 			await this.explorerService.applyBulkEdit(resourceFileEdits, options);
 		} catch (error) {
+
 			// Conflict
 			if ((<FileOperationError>error).fileOperationResult === FileOperationResult.FILE_MOVE_CONFLICT) {
 
@@ -1497,20 +1109,17 @@ export class FileDragAndDrop implements ITreeDragAndDrop<ExplorerItem> {
 					}
 				}
 
-				const confirm = getMultipleFilesOverwriteConfirm(overwrites);
 				// Move with overwrite if the user confirms
+				const confirm = getMultipleFilesOverwriteConfirm(overwrites);
 				const { confirmed } = await this.dialogService.confirm(confirm);
 				if (confirmed) {
-					try {
-						await this.explorerService.applyBulkEdit(resourceFileEdits.map(re => new ResourceFileEdit(re.oldResource, re.newResource, { overwrite: true })), options);
-					} catch (error) {
-						this.notificationService.error(error);
-					}
+					await this.explorerService.applyBulkEdit(resourceFileEdits.map(re => new ResourceFileEdit(re.oldResource, re.newResource, { overwrite: true })), options);
 				}
 			}
-			// Any other error
+
+			// Any other error: bubble up
 			else {
-				this.notificationService.error(error);
+				throw error;
 			}
 		}
 	}
